Refactor loudness task boundaries
